Connect to Notification Hub
When user request to join Group, notification will be sent to all the
group's leaders

+++ WIP/Source/iVolunteer/iVolunteer/Hubs/ChatHub.cs
@@ -43,19 +43,26 @@ namespace iVolunteer.Hubs
         /// <returns></returns>
         public override Task OnDisconnected(bool stopCalled)
         {
-            string connectionID = Context.ConnectionId;
-            try
+            if (stopCalled)
             {
-                SQL_HubConnection_DAO connDAO = new SQL_HubConnection_DAO();
-                // Who has disconnected???
-                string userID = connDAO.Get_UserID(connectionID);
-                connDAO.Delete_Connection(connectionID);
+                string connectionID = Context.ConnectionId;
+                try
+                {
+                    SQL_HubConnection_DAO connDAO = new SQL_HubConnection_DAO();
+                    // Who has disconnected???
+                    string userID = connDAO.Get_UserID(connectionID);
+                    connDAO.Delete_Connection(connectionID);
 
-                Clients.Others.userHasDisconnected(userID);
+                    Clients.Others.userHasDisconnected(userID);
+                }
+                catch
+                {
+                    throw;
+                }
             }
-            catch
+            else
             {
-                throw;
+                
             }
             return base.OnDisconnected(stopCalled);
         }
